+++ robocorp-code/codegen/commands.py
@@ -168,6 +168,22 @@ COMMANDS = [
         icon={"light": "images/light/debug.svg", "dark": "images/dark/debug.svg"},
     ),
     Command(
+        "robocorp.runRobocorpsPythonTask",
+        "Run Robocorp's Python Task",
+        add_to_package_json=True,
+        server_handled=False,
+        hide_from_command_palette=True,
+        icon={"light": "images/light/run.svg", "dark": "images/dark/run.svg"},
+    ),
+    Command(
+        "robocorp.debugRobocorpsPythonTask",
+        "Debug Robocorp's Python Task",
+        add_to_package_json=True,
+        server_handled=False,
+        hide_from_command_palette=True,
+        icon={"light": "images/light/debug.svg", "dark": "images/dark/debug.svg"},
+    ),
+    Command(
         "robocorp.saveInDiskLRU",
         "Saves some data in an LRU in the disk",
         add_to_package_json=False,