+++ robocorp-code/codegen/settings.py
@@ -104,6 +104,12 @@ SETTINGS = [
         "Specifies the timeout in minutes for the token generated to access the vault when a launch is made. Note: max 60, min 5, additional timeout may be added internally to reuse the token in future runs (please use Robocorp Control Room or Robocorp Assistant for longer runs).",
         setting_type="number",
     ),
+    Setting(
+        "robocorp.codeLens.roboLaunch",
+        True,
+        "Specifies whether the 'Run Task' and 'Debug Task' code lenses should be shown.",
+        setting_type="boolean",
+    ),
 ]